requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
defusedxml==0.7.1
Django==4.0.4
django-cors-headers==3.12.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
lxml==4.8.0
pytz==2022.1
regex==2022.4.24
sqlparse==0.4.2
websocket-client==1.3.2
